poke-mcp-server/pyproject.toml
[project]
name = "poke-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "fastmcp>=2.13.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
